refactor: rename `config`
