+++ lib/util/launcher_info.dart
@@ -4,7 +4,7 @@ import 'package:feedback/feedback.dart';
 import 'package:flutter/material.dart';
 import 'package:path/path.dart';
 import 'package:pub_semver/pub_semver.dart';
-import 'package:rpmlauncher/util/Config.dart';
+import 'package:rpmlauncher/util/config.dart';
 import 'package:rpmlauncher/util/updater.dart';
 import 'package:rpmlauncher/util/I18n.dart';
 import 'package:rpmtw_dart_common_library/rpmtw_dart_common_library.dart';
