+++ test/widget/widget_test.dart
@@ -7,7 +7,7 @@ import 'package:rpmlauncher/model/account/Account.dart';
 import 'package:rpmlauncher/model/Game/GameLogs.dart';
 import 'package:rpmlauncher/model/IO/Properties.dart';
 import 'package:rpmlauncher/screen/Account.dart';
-import 'package:rpmlauncher/util/Config.dart';
+import 'package:rpmlauncher/util/config.dart';
 import 'package:rpmlauncher/util/data.dart';
 import 'package:rpmlauncher/util/I18n.dart';
 import 'package:rpmlauncher/view/RowScrollView.dart';
